script to generate spreadsheets
python script generate_peer_review will generate spreadsheets for the
peer review exercise for each member of the class.

The names of the students should be stored in a csv file in format

Full name, team, student ID

a sample teams.csv is provided.

The template spreadsheet contains the 6 questions under which students
provide marks for the team mates, a cell for comments, and a total
score. The final spreadsheet is heavily protected so that students must
follow the rules for allocating marks.

The instructions to students for completing the spreadsheet are located
in the text file "instructions.txt".

execute with python3 generate_peer_review.py.